Transformação de Dados/requirements.txt
pandas>=1.3.0
tabula-py>=2.5.0
PyPDF2>=2.0.0
python-dotenv>=0.19.0